+++ android/app/src/main/res/layout/fragment_my_page.xml
@@ -7,7 +7,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="?attr/colorSurface"
-        tools:context=".ui.MyPageFragment">
+        tools:context=".ui.mypage.MyPageFragment">
 
         <com.google.android.material.appbar.AppBarLayout
             android:id="@+id/abl_signup_topbar"
