+++ android/app/src/main/res/layout/item_my_page_list.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="settingInfo"
-            type="app.priceguard.ui.SettingItem" />
+            type="app.priceguard.ui.mypage.SettingItem" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
